Upgrade to laravel 5.8.x.

+++ database/migrations/2014_01_07_073615_create_tagged_table.php
@@ -13,7 +13,7 @@ class CreateTaggedTable extends Migration
             if (config('tagging.primary_keys_type') == 'string') {
                 $table->string('taggable_id', 36)->index();
             } else {
-                $table->integer('taggable_id')->unsigned()->index();
+                $table->bigInteger('taggable_id')->unsigned()->index();
             }
             $table->string('taggable_type', 125)->index();
             $table->string('tag_name', 125);
