+++ database/migrations/2018_02_23_122334_create_articles_table.php
@@ -14,22 +14,22 @@ class CreateArticlesTable extends Migration
     public function up()
     {
         Schema::create('articles', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('title');
             $table->longText('text');
             $table->string('source')->nullable();
             $table->text('description')->nullable();
             $table->string('slug')->nullable();
-            $table->integer('category_id');
-            $table->integer('user_id')->nullable();
+            $table->bigInteger('category_id');
+            $table->bigInteger('user_id')->nullable();
             $table->boolean('enabled');
             $table->timestamps();
             $table->softDeletes();
         });
         Schema::create('article_comments', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('product_id');
-            $table->integer('user_id');
+            $table->bigInteger('product_id');
+            $table->bigInteger('user_id');
             $table->text('question')->nullable();
             $table->text('answer')->nullable();
             $table->boolean('enabled');
