+++ database/migrations/2018_02_23_172722_create_transactions_table.php
@@ -14,14 +14,14 @@ class CreateTransactionsTable extends Migration
     public function up()
     {
         Schema::create('transactions', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('invoice_id')->nullable();
+            $table->bigIncrements('id');
+            $table->bigInteger('invoice_id')->nullable();
             $table->unsignedBigInteger('gateway_transaction_id')->nullable();
             $table->unsignedBigInteger('gateway')->nullable();
-            $table->integer('account_id')->nullable();
-            $table->integer('project_id')->nullable();
-            $table->integer('user_id')->nullable();
-            $table->integer('category_id')->nullable();
+            $table->bigInteger('account_id')->nullable();
+            $table->bigInteger('project_id')->nullable();
+            $table->bigInteger('user_id')->nullable();
+            $table->bigInteger('category_id')->nullable();
             $table->string('currency_code')->default(config('platform.currency'))->nullable();
             $table->double('currency_rate', 15, 8)->nullable();
             $table->decimal('amount', 15, 4);
